refactor(content-fetcher): extract helpers to reduce file from 391 to 67 lines
Split content-fetcher.js into focused modules:
- content-fetcher-html.js (103 lines): HTML parsing, text extraction
- content-fetcher-browser.js (95 lines): Playwright, Google Cache
- content-fetcher-http.js (88 lines): HTTP fetch with retries
- content-fetcher.js (67 lines): main orchestration (was 391)

All functions are now under 20 lines.
